Fix format strings in searchcorespi, fastlib, storageframework, vespalib and streamingvisitors module.

+++ vespalib/src/vespa/vespalib/trace/trace.cpp
@@ -54,7 +54,7 @@ Trace::trace(uint32_t level, const string &note, bool addTime)
     if (addTime) {
         struct timeval tv;
         gettimeofday(&tv, NULL);
-        _root.addChild(make_string("[%ld.%06ld] %s", tv.tv_sec, tv.tv_usec, note.c_str()));
+        _root.addChild(make_string("[%ld.%06ld] %s", tv.tv_sec, static_cast<long>(tv.tv_usec), note.c_str()));
     } else {
         _root.addChild(note);
     }
